created feature folder & moved commands in there
