+++ src/features/commands/mods.js
@@ -5,9 +5,9 @@
 
 
 const { EmbedBuilder, SlashCommandBuilder, ActionRowBuilder, ButtonBuilder, ButtonStyle } = require('discord.js');
-const ModsData = require("../data/mods.js");
-const SystemUtils = require("../utils/SystemUtils.js");
-const config = require("../config/config.json")
+const ModsData = require("../../data/mods.js");
+const SystemUtils = require("../../utils/SystemUtils.js");
+const config = require("../../config/config.json")
 
 
 const subcommands = {
