+++ src/features/commands/version.js
@@ -4,9 +4,9 @@
 //
 
 const { EmbedBuilder, SlashCommandBuilder } = require('discord.js');
-const MainMenuData = require("../data/main_menu.js");
-const SystemUtils = require("../utils/SystemUtils.js");
-const config = require("../config/config.json")
+const MainMenuData = require("../../data/main_menu.js");
+const SystemUtils = require("../../utils/SystemUtils.js");
+const config = require("../../config/config.json")
 
 const subcommands = {
     get: { name: "get", function: handleGetCommand, permission: false },
